+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## 1.1.0 July 24, 2023
 
+- Enh #617: Add debug collector for yiisoft/yii-debug (@xepozz)
 - Chg #722: Remove legacy array syntax for typecast. Use `Param` instead (@terabytesoftw)
 - Chg #724: Typecast refactoring (@Tigrov)
 - Chg #728: Refactor `AbstractSchema::getColumnPhpType()` (@Tigrov)